Update renamed files in source code
